Rename useStore to useConnect
